back-end/pyproject.toml
[tool.poetry]
name = "squad-sixteen"
version = "0.1.0"
description = "Projeto universitário em colaboração com a JotaNunes, tendo como objetivo a construção de uma plataforma de cursos"
authors = ["Raphael Santana <r.laurindo.santanapro@gmail.com>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.8"}
sqlalchemy = "^2.0.40"
bcrypt = "^4.3.0"
psycopg2 = "^2.9.9"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
passlib = "^1.7.4"
python-jose = "^3.3.0"  # para `jose`
pyjwt = "^2.8.0"
pandas = "^2.2.3"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
